+++ pytorch_lightning/loops/base.py
@@ -33,12 +33,6 @@ class Loop(ABC):
 
     .. codeblock:: python
 
-        if is_restarting:
-            restore()
-            is_restarting = False
-        else:
-            reset()
-
         on_run_start()
 
         while not done:
@@ -52,15 +46,15 @@ class Loop(ABC):
     def __init__(self) -> None:
         self.iteration_count: int = 0
         self.trainer: Optional['pl.Trainer'] = None
-        self._is_restarting = False
+        self._restarting = False
 
     @property
-    def is_restarting(self) -> bool:
-        return self._is_restarting
+    def restarting(self) -> bool:
+        return self._restarting
 
-    @is_restarting.setter
-    def is_restarting(self, is_restarting: bool) -> None:
-        self._is_restarting = is_restarting
+    @restarting.setter
+    def restarting(self, restarting: bool) -> None:
+        self._restarting = restarting
 
     @property
     @abstractmethod
@@ -102,9 +96,9 @@ class Loop(ABC):
         if self.skip:
             return self.on_skip()
 
-        if self.is_restarting:
+        if self.restarting:
             self.restore()
-            self.is_restarting = False
+            self.restarting = False
         else:
             self.reset()